requirements.txt
Pyrebase4==4.8.0
streamlit==1.41.1
pandas==2.2.1